P13 Move dropdown component to separate file.
